sunxi-disp: ignore invalid arguments for FBIOPUTCMAP
Just do nothing and return 0 for the unsupported palette indexes
instead of returning -EINVAL. The vesafb and uvesafb drivers do
the same.

Otherwise xf86-video-fbdev or anything else derived from it writes
a lot of "FBIOPUTCMAP: Invalid argument" spam to /var/log/Xorg.0.log
when the X server starts. And apparently also when the screensaver
and/or screen blanking/unblanking kicks in. These error messages
are scaring the users and also distracting attention when doing
real troubleshooting. Not to mention slowing down the X server
startup and wasting extra disk space for storing logs.

Yes, it is ugly to ignore errors. And the right solution would
be to fix the userland code not to use this ioctl with wrong
arguments. But frankly speaking, I don't want to touch the
prehistoric paletted formats in the X server as long as we don't
have a real use case where they can be tested.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>