sun4i-ts: Simplify code significantly
The sun4i-ts controller is capable of detecting a second touch, but when a
second touch is present then the accuracy becomes so bad the reported touch
location is not useable.

The android driver contains some complicated heuristics using the aprox.
distance between the 2 touches to see if the user is making a pinch open /
close movement, and then reports emulated multi-touch events around the last
single touch coordinate (as the dual-touch coordinates are worthless).

Note that this pinch emulation has never worked as the implementation of
the evdev mt protocol inside the android code is wrong, more over it is
wrong because there is a lot more to multi-touch then just pinch detection,
and reporting being mt capable when we're really not will confuse apps.

I'm currently writing a new rtp driver from scratch for upstream, I've decided
to start with gutting all the heuristics from the 3.4 driver making it work
the same way as I intend the upstream driver to work, so I can test things
on a kernel where I have a working lcd driver.

So this commit removes all the broken heuristics. As a bonus after this patch
the tslib.patch people have been using also is no longer needed:
https://raw.github.com/webconn/OLINUXINO/master/SOFTWARE/A13/TOUCHSCREEN/tslib.patch

Signed-off-by: Hans de Goede <hdegoede@redhat.com>