sunxi: Enable ZONE_DMA for sun4i and sun5i (dma_zone_size=256MB)
Previously only sun7i had ZONE_DMA enabled, which is originating
from the Allwinner linux-3.3 code drop. But there should be no
reason to have it different for sun4i and sun5i.

The practical effect of setting the dma_zone_size is that it
affects the placement of the CMA area in physical RAM (if CMA
is enabled in the kernel). And we want to have it placed in the
first 256MB of physical RAM because the Cedar VPU is going to
be unhappy otherwise. For example, the encoding of the address
in MACC_MPEG_VLD_ADDR register is a bit tricky for anything that
needs more than 28 bits:
    http://linux-sunxi.org/VE_Register_guide#MACC_MPEG_VLD_ADDR
The low 4 bits might be supposed to store the high part of the
address, but in practice the video playback is currently broken
(in the proprietary libve and in the open source libvdpau-sunxi)
if Cedar memory buffer is allocated at a physical address
above 256MB.

TL;DR this workarounds Cedar VPU issues in the case if CMA is
enabled.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>
Acked-by: Hans de Goede <hdegoede@redhat.com>